Resolved merge conflicts
