+++ src/layouts/BaseLayout/Header/AuthButton.tsx
@@ -1,14 +1,14 @@
 import { Link } from 'react-router-dom';
 
-import ROUTES from '@/constants/routes';
-import { useAuth } from '@/context/auth/useAuth';
+import { useAuth } from '@/provider/auth/useAuth';
+import { ROUTER_PATH } from '@/routes/path';
 
 import { Button } from '@/components/ui/Button';
 
 export const AuthButton = () => {
   const { isLoggedIn } = useAuth();
 
-  const linkTo = isLoggedIn ? ROUTES.MY_ACCOUNT : ROUTES.LOGIN;
+  const linkTo = isLoggedIn ? ROUTER_PATH.MY_ACCOUNT : ROUTER_PATH.LOGIN;
   const text = isLoggedIn ? '내 계정' : '로그인';
 
   return (
