+++ src/pages/NotFound.tsx
@@ -3,7 +3,7 @@ import { css } from '@emotion/react';
 import { Link } from 'react-router-dom';
 
 import { API_ERROR_MESSAGES } from '@/constants/errorMessage';
-import ROUTES from '@/constants/routes';
+import { ROUTER_PATH } from '@/routes/path';
 
 import { Button } from '@/components/ui/Button';
 import { Container } from '@/components/ui/Layout/Container';
@@ -25,7 +25,7 @@ const NotFound = () => {
       css={containerStyle}
     >
       <h1 css={titleStyle}>{API_ERROR_MESSAGES.PAGE_NOT_FOUND}</h1>
-      <Link to={ROUTES.HOME}>
+      <Link to={ROUTER_PATH.HOME}>
         <Button theme="outline" css={{ width: '12rem' }}>
           메인 페이지로 돌아가기
         </Button>
