+++ src/routes/path.ts
@@ -1,8 +1,6 @@
-const ROUTES = {
+export const ROUTER_PATH = {
   HOME: '/',
   THEME: '/theme/:themeKey',
   LOGIN: '/login',
   MY_ACCOUNT: '/my-account',
 };
-
-export default ROUTES;
