--- src/styles/variants/breakpoint.ts
@@ -1,7 +1,0 @@
-export const breakpoint = {
-  initial: '0px',
-  xs: '520px',
-  sm: '768px',
-  md: '1024px',
-  lg: '1280px',
-};
