--- src/styles/variants/theme.ts
@@ -1,35 +1,0 @@
-export const colors = {
-  yellow: {
-    100: '#FEEB00',
-    200: '#F2DA00',
-  },
-  gray: {
-    100: '#F2F2F2',
-    200: '#DCDCDC',
-    300: '#D4D4D4',
-    400: '#404040',
-    500: '#262626',
-  },
-  blue: {
-    100: '#E6F1FF',
-    200: '#D6E5FF',
-    300: '#70A4F9',
-    400: '#4584E9',
-  },
-  black: '#000000',
-  white: '#FFFFFF',
-  blackOpacity: {
-    '20': 'rgba(0, 0, 0, 0.2)',
-    '40': 'rgba(0, 0, 0, 0.4)',
-    '60': 'rgba(0, 0, 0, 0.6)',
-    '80': 'rgba(0, 0, 0, 0.8)',
-  },
-  whiteOpacity: {
-    '20': 'rgba(255, 255, 255, 0.2)',
-    '40': 'rgba(255, 255, 255, 0.4)',
-    '60': 'rgba(255, 255, 255, 0.6)',
-    '80': 'rgba(255, 255, 255, 0.8)',
-  },
-  pink: '#F99393',
-  red: '#FF0000',
-};
